Python Proj2/tesser/requirements.txt
pdf2image
pytesseract
Pillow
Flask
python-docx
